More code reformat, moving some things around, implementing routes, handlers
