+++ src/main/java/net/staticsnow/nexus/repository/apt/internal/AptWritePolicySelector.java
@@ -19,19 +19,22 @@ import org.sonatype.nexus.repository.storage.Asset;
 import org.sonatype.nexus.repository.storage.WritePolicy;
 import org.sonatype.nexus.repository.storage.WritePolicySelector;
 
-public class AptWritePolicySelector implements WritePolicySelector {
+public class AptWritePolicySelector
+    implements WritePolicySelector
+{
 
-	@Override
-	public WritePolicy select(Asset asset, WritePolicy configured) {
-		if (WritePolicy.ALLOW_ONCE == configured) {
-			String name = asset.name();
-			if (name.endsWith(".deb")) {
-				return WritePolicy.ALLOW_ONCE;
-			} else {
-				return WritePolicy.ALLOW;
-			}
-		}
-		return configured;
-	}
+  @Override
+  public WritePolicy select(Asset asset, WritePolicy configured) {
+    if (WritePolicy.ALLOW_ONCE == configured) {
+      String name = asset.name();
+      if (name.endsWith(".deb")) {
+        return WritePolicy.ALLOW_ONCE;
+      }
+      else {
+        return WritePolicy.ALLOW;
+      }
+    }
+    return configured;
+  }
 
 }
